modules: implement rate_limit scheduler
This scheduler prevents from sending more than X actions per second,
or processing more than given volume of data.
It is based on the token bucket filter algorithm.

This can be configured using a 'rate_limit' sub-block in policy parameters:
	rate_limit {
		max_count = 10000;
		max_size = 10GB;
		period_ms = 10000;
	}

This patch defines a scheduler that is loaded with the "common" module,
so it is available as 'common.rate_limit'.

Add test 237 to check its behavior.

Add example configurations for lhsm_archive and checksum policies.

Change-Id: I37c0bfac298d978f03cd10d0103ae2c00cbb0166
Signed-off-by: Henri Doreau <henri.doreau@cea.fr>
Signed-off-by: Thomas Leibovici <thomas.leibovici@cea.fr>