Update patch set 1
Patch Set 1: -Code-Review

I hesitate about this patch.
On the one hand, it brings a security for most situations.
But on the other hand, we could imagine policies where we want to remove entries from the backend whereas they stil exist in the filesystem (for example for modified entries).

I suggest you does this check in "lhsm_remove" function, so this function becomes safe for the common usage.
And create a new "lhsm_force_remove" function for those who would like to remove existing entries on purpose.

Patch-set: 1