Update patch set 1
Patch Set 1:

this might not be the best way to handle the issue, I agree. there could be other ways to handle this case.
if a file exists but SOFT_RM time for the entry expired, how could this be handled ?
- extend by another 60 days the file remove in SOFT_RM table ?
- recall the file and call hsm_remove to remove entry from backend ?
other options ?
this case need to be handled to avoid soft_rm process from exiting because of error counts.

Patch-set: 1
Reviewer: Tack <1009999@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0