Update patch set 1
Patch Set 1:

The main points are:
* It is licit to run hsm_remove on an existing entry
* It is abnormal an entry is in the SOFTRM table if it still exists

Definitely, the fix should be to avoid existing entries to be moved to SOFTRM table. This should not happen. If this happens, this is a bug.

The test case you propose highlight a case where this happens: moving an archived entry to an ignored tree. In this case robinhood doesn't scan the tree and consider the entries as deleted.

The problem in that case is that entries are moved to SOFTRM by ListMgr_MassSoftRemove(), and I dislike the design of doing filesystems calls from the listmgr layer... Or perharps by providing a callback function...?

Patch-set: 1
Reviewer: Thomas Leibovici <1003874@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0