changelog reader: detailed changelog stats per MDT
Enrich the changelog stats and maintain them for each changelog reader (each
MDT).
Now it maintains the last processed record id, record timestamp and the time
of processing for the 4 steps of the robinhood workflow:
- reading the changelog from lustre
- pushing it to the entry processing pipeline (can be delayed due to the
  changelog batching mechanism),
- committing the changelog information to the database
- clearing the changelog from lustre

These stats are displayed in the logs with other STATS and are also stored in
the DB so they are available using the "rbh-report -a" command.

Some variable in the DB have been renamed, removed or added. The
cleaning/conversion process is done once when the following situation is
detected:
when a changelog reader starts and tries to retrieve the last committed record
id from the VARS table, if it doesn't find the new variable name and find the
old one, it cleans old stat variables.

Change-Id: I0e76a95f3f9bc4406c7d3cd3c2cb94c1e59d7177