changelog reader: regular update of last committed record id
Regularly update the last committed record id in the DB to avoid
reading again old changelog records after robinhood restarts.

This may be expensive to update it after each DB commit,
and this would require a big change in the ListMgr API to update it
together with records. So we update this id at regular interval.

This change introduces 2 config parameters to the changelog reader
configuration to control this update frequency:
"commit_update_max_delta" and "commit_update_max_delay".

Add test 124 to check this mechanism.

Change-Id: If6877593ac6e903139fb00662d37919d05fd38c8