FS scan: restrict scan to several directories
Admin can specify several "scan_only" parameters in the FS_Scan config block.
In this case, the scan is restricted to these sub-directories instead of
scanning the whole filesystem.

In particular, this is convenient to manage static partitioning with DNE
with 1 robinhood instance per MDT. In this case, we can restrict the
scan of secondary MDTs only to the directories they manage.

Add test 126 to verify this feature.

Change-Id: Ic09306413b3ad96c77938998e4378ed4330feb3a