entry processor: keep stripe info for matching
When stripe information was up-to-date in DB, it was cleared from the entry
attributes during the processing, thus resulting in wrong matching of
stripe-related criteria (pool, ost_index, ...).
This was fixed by keeping the stripe info until the DB operation,
and keeping a boolean that indicates if the stripe information is up-to-date
in DB.

Completed test 614 to check pool matching, including the "no pool" case.
Test both 'scan' and 'readlog' cases.

Change-Id: Id1df382e4e5eaf78c5795986512bbb6f2f028aab