policy: get entry status even if it needed in the scope
Fixes the following issue: policy scope was matched to determine
if we must call getstatus() on it. But there was a chicken egg issue
in this case, as the scope could not be matched because the status is
missing at this point...
So it makes scope matching tolerant to missing status.

Change-Id: I275bc89db79e05e2ceb0c2a20a9d364861ad1432