Update patch set 1
Patch Set 1:

Break tests 112,115,404 on POSIX FS.

Actually the issue is triggered when the FS mount point contains a special character (e.g. dot in /mnt/rbh.test).
The problem is when we check if the given path criteria is under FS root. Is there a C function (like fnmatch) to match a string with such regexp?

Patch-set: 1
Label: Verified=0