Update patch set 1
Patch Set 1:

See "%exclude" directives. This excludes the shook module (.so) from the main package "robinhood-lustre" by packaging it in a different RPM. Having this module in the main package make it depend on shook libraries (automatic dependency detection by rpmbuild), what we don't want. With this patch, only the robinhood-mod-shook has a depencency on shook libraries.

Patch-set: 1
Label: Code-Review=0