Update patch set 7
Patch Set 7: Code-Review+1

(1 comment)

OK for the code contents. We can use it as is (as a local patch for now?).
autotool improvements are possible (I think the name of the option), we can do them later(R) when we wish to land this upstream.

Patch-set: 7
Reviewer: Thomas Leibovici <1003874@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1