listmgr: stop retrying SQL requests on SIGTERM
Stop Robinhood on signal during DB retries

Consider the situation: Robinhood lost connection to MySQL server
and attempts to retry some request. Even if a TERM signal is received,
Robinhood will attempt to retry infinitely.

Adding a boolean flag to set in the signal handler and checking it in
SQL request handlers solves the issue.

Change-Id: Iccae3c50dd67558f190c2f5d158288473de16531