Update patch set 2
Patch Set 2:

> Here they are:

Thanks!

> rbh_find_printf.c: In function ‘printf_entry’:
> rbh_find_printf.c:602: error: #pragma GCC diagnostic not allowed inside functions
> rbh_find_printf.c:603: error: #pragma GCC diagnostic not allowed inside functions
> rbh_find_printf.c:605: error: #pragma GCC diagnostic not allowed inside functions

Oh. That's going to be annoying, we could try moving the #pragma to disable the warning around the whole function but that's kind of a shame.

Although I find it rather odd that gcc/clang complain about this one but not about the other built format strings later on, it's a bit disingenuous...
Well, the only alternative I can see would be to manually deduplicate the double-% signs in the 0th case of the switch as it's the only reason we use printf here, but it feels like losing to compilers so I don't like it :)

So trending towards just moving the pramga out.. Oh, I guess pragmas are processed after m4 so we could put them into #if __GNUC__ >= 7... Not sure what's better.

Patch-set: 2
Reviewer: Dominique Martinet <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>