Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Overall this is a good idea, but not really compatible with the direction where we're going: we've noticed that running many policies in parallel will slow down the rest of robinhood (changelog reader or scan) as it does many forks, so we're planning on making the service more flexible to run multiple instances of robinhood at the same time.

This would not work with a convention that blindly Requires=%i.mount as, well, there can be only one ;)

Do you know why remote-fs is too slow for you?


Meanwhile, I encourage you to put that extra require in a systemd override local to your site.

Patch-set: 1
Reviewer: Dominique Martinet <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1