Update patch set 1
Patch Set 1:

> It's not so much performance, but convenience. We run RobinHood instances for several filesystems on the same host. Having systemd remember dependencies is easier and less error-prone if we have to temporarily take down an fs than having to do so ourselves. More importantly, though, depending on how a certain filesystem is started/mounted, it may be out of scope for remote-fs.target, whereas %i.mount (with an adequate service name) produces the correct dependency.

I see.
If just failing the service start is fine, check_mounted would be enough, but to actively make systemd mount it for you when you start the service the only thing I can come up with that'd work for everyone would be an ExecStartPre command that takes in variables from the sysconfig file.

This is quite ugly, but I do not see any other way of pulling in something that's not directly related to the suffix...
The arbitrary command could be as simple as 'systemctl start mnt-foo-bar.mount', I think that will properly wait until the filesystem is actually mounted if defined in fstab, do nothing if already mounted, and fail if the filesystem isn't configured.


That being said, I'm honestly not convinced this is better than drop-ins - we could just add a comment that says feel free to add require=x.mount in an override conf?
That'd also be flexible enough to have two robinhood@{x,y} require the same mount unit yet have robinhood@z require some other one

Patch-set: 1
Reviewer: Dominique Martinet <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>