tests: use a smaller/simpler test set for test 113
Test 113 was broken when /bin is a symlink to /usr/bin.
Moreover, the test was disturbed by hardlinks in this directory.
Additionally, this directory could be quite large, thus taking
a significant time to run the test.

Now copy robinhood test dir instead of /bin, as it contains a simpler
and smaller set of entries.

Change-Id: Id6ab540499795a0217673dfa1fb30841b1789d45