Lustre: workaround to support PFL
This initial implementation is a workaround to provide minimal
support of Progressive File Layouts.

It only stores and handles the striping information of the last
initialized component of a layout, which is likely the one suited
for the current file size.

Also verified compliance with DoM and HSM.

A better support of PFL will require a DB schema change,
and a deep change of stripe handling in robinhood.

Change-Id: Id6d1c7068f4ccf72e4268074df5dc0c364ee4f8b