policy_triggers: add sanity check about statfs structure
On some buggy systems, statfs() reports no error but the returned
structure is inconsistent or is filled with zeros...
This sanity check was already done for 'df' and 'lfs df' based triggers,
but not for triggers based on user and group usage (in this case, df is
still used as the reference to get the total filesystem volume).
This bug was last triggered on Lustre 2.12.3.

Change-Id: Icfac778b26cd282774236c6abf07dbf3c4438da2
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/487089
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Tested-by: Thomas Leibovici <thomas.leibovici@cea.fr>