triggers: fix optimization for policy runs per OST
The implemented optimization was buggy for triggers running on multiple
targets like OSTs, pools...
Indeed, for sorted policies it stored the time of the last eligible
entry to restart from this value at the next policy run. It's OK if
the list is the same for each policy run, which is not the case if the
policy runs on multiple OSTs or pools.

Change-Id: Ie38be057c7ad0c805fd0c75bb3d2dc6cfbed6db4
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/512106
Reviewed-by: Dominique Martinet <asmadeus@codewreck.org>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Tested-by: Thomas Leibovici <thomas.leibovici@cea.fr>