policy: update db only for refreshed file attrs
When pre_sched_match and post_sched_match are set to "none or
"cache_only" (no stat is done on the entry), we don't have to update
file attrs on the database after or before the policy.

This patch is an optimization to relax the load on the robinhood
database.

Change-Id: I3c5fc8735738adfc2597d6091b69904791bcc42a
Signed-off-by: Etienne AUJAMES <eaujames@ddn.com>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/541697
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Tested-by: Thomas Leibovici <thomas.leibovici@cea.fr>