pkg: Add check for php-json
php-json package is used by webgui. While
the build would pass if php-json package is
missing. It may not render the graphs at all.

This patch adds check for php-json. During the
time of configure. It does not change workflow
or alters functions. If php-json is missing it
will throw an "warning" under config.log so
user can review it post running and add that if
necessary. Hard requirement of php-json was not
selected as not all users would like to use
webgui. Therefore the changes under robinhood.spec.in
("BuildRequires:") was not chosen.

Change-Id: I45c3aeccdd4983a5ae9935dc383b55ced9ffbc5e
Signed-off-by: Arshad Hussain <arshad.hussain@aeoncomputing.com>