test 301: rewrite the data generation loop
The loop was creating files from $empty_vol to $mb_h_threshold.
This issue was that it created files named file.$i where $i goes from
$empty_vol to $mb_h_threshold. But nothing guaranties that some of the
files that constitutes the volume $empty_vol on the OSTs are not named
like this. In this case, the loop would simply rewrite a file that
already exists thus not increasing the amount of data on the OSTs.

Also, flush the page and lru cache after the creation of the files to
ensure that everything is committed to disk. This seems to improve the
reliability of the test.

Change-Id: Id11f1b4d00d2eb8b435e2604327ecfbf66bbc7f1
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/1175761
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Tested-by: Guillaume Courrier <guillaume.courrier@cea.fr>