test 115: do not clear changelogs for LUSTRE_HSM flavor
This test checks whether the backup module can handle a missing parent
path in the database when doing an undelete. This test doesn't seem to
work for the LUSTRE_HSM and TPMFS configurations. The issue is that
rbh-undelete won't work if the parent is not in the DB but it was
removed by the test. With the backup module, if the parent is not
present, the file will be restored in the backup directory. But for
other flavors rbh-undelete will simply fail.

For other flavors, we need to let the changelog reader read the
changelogs so that it can detect that the parent is missing from the DB
and restore it.

Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Change-Id: I9c131210ddc9d0e17678fbf2605a9a98c915c67c
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/1175762
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>