tests: fix test 102
Rename changelog records were not all processed by the daemon when
killed in the test loop. This resulted in leaving the wrong name for the
entry in the DB ('file.2' instead of 'file').
When the later HSM changelog record was processed, this resulted
in having 2 sames for the file in the DB, as if it was 2 hardlinks.
As a result, this ruined the logic of testing the object's path_update.

To fix this, I added a final "--readlog --once" command after killing
the changelog reader, to ensure all records are read.

Change-Id: Ie135731f98e0aedf45ef891ab5f822c89b492a85
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/1200334
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Tested-by: Thomas Leibovici <thomas.leibovici@cea.fr>