tests: improve tests based on df
'df' update latency causes test failures because we rely on its value
to trigger policies, and to compute the number of actions to be
executed.
Add functions to make sure the df value is updated before using its
value for tests.
This does not completely eliminate the df latency issue, but reduces its
frequency and impact.

Change-Id: Icc6448257405505da3c418a0a644fc4651963729
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/1200788
Tested-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>