rbh_report: count all entries in top-users if --by-count is specified.
When reporting top users with --by-count, the count of inodes
is important and should be the same as what we can observe on the
filesystem.
Directories and symlinks were not taken into account when reporting top
user files count.

This patch no longer filter files when --by-count is specified.

Note that this behaviour can be confusing, regarding how we want to
    sort, the filecount will be different.

Change-Id: Idef43fecba3e8911757c0bc151f93c25d9d32103
Signed-off-by: Aurélien Cedeyn <aurelien.cedeyn@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/cea-hpc/robinhood/+/1204023
Tested-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>